Add temporary invulnerability on teleport
This change introduces a feature that makes players temporarily invulnerable after teleporting. This is to prevent players from taking damage from falling or suffocation when the destination chunk has not loaded yet.

The default invulnerable time now is stored in the config JSON file for ease of customisation.
Please remove the old config/HomeLink/HomeLink.json file to allow the generation of an updated version.